change targets to <PLATFORM>_<TEST> + disallow _ in test name;
before we had run-<PLATFORM>--<TEST> - the double-dash was needed
because the platform names contain - and this makes splitting the
platform and the target again from the target name too tedious.
I changed it to run-<PLATFORM>_<TEST> now, but that means we can't have
_ in both platform and test. There are some tests with _, but I think
that's bad practice anyway, so I'll rename them.